pubsub_discoverer_v1/backend/app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Bruce Lau <bruce.lau@telus.com>"]

[tool.poetry.dependencies]
python = "^3.7"
uvicorn = "0.20.0"
fastapi = "0.91.0"
python-multipart = "^0.0.5"
email-validator = "^1.0.5"
requests = "^2.23.0"
passlib = {extras = ["bcrypt"], version = "^1.7.2"}
tenacity = "^6.1.0"
pydantic = "^1.4"
raven = "^6.10.0"
gunicorn = "^20.0.4"
jinja2 = "^2.11.2"

pytest = "^5.4.1"
python-jose = {extras = ["cryptography"], version = "^3.1.0"}
starlette-exporter = "^0.4.1"
starlette-context = "^0.2.3"
aiofiles = "^0.5.0"
cryptography = "3.3.0"
PyJWT = "^2.3.0"

tinaa-logger = ">=3.0,<4.0"
anyio = "^3.6.2"
avro = "^1.11.1"
avrotool = "^0.0.2"
fastavro = "^1.7.3"
greenlet = "^2.0.2"
h11 = "^0.14.0"
idna = "^3.4"
packaging = "^23.1"
paramiko = "^3.1.0"
peppercorn = "^0.6"
python-snappy = "^0.6.1"
rec-avro = "^0.0.4"
sniffio = "^1.3.0"
sqlalchemy = "^2.0.10"
tinaa-utils = "^2.0.2"
typing-extensions = "4.4.0"

[tool.poetry.dev-dependencies]
mypy = "^0.770"
black = "^19.10b0"
isort = "^4.3.21"
autoflake = "^1.3.1"
flake8 = "^3.7.9"
pytest = "^5.4.1"

pytest-cov = "^2.8.1"
watchdog = "^2.1.3"
argh = "^0.26.2"

[[tool.poetry.source]]
name = 'tinaa'
url = 'https://gitlab.tinaa.teluslabs.net/api/v4/projects/1029/packages/pypi/simple'
secondary = true

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 88
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
